+++ src/app/account/account.component.html
@@ -1,4 +1,4 @@
-<!-- <div class="mainCard">
+ <!-- <div class="mainCard">
   <div class="mainCard__title">List with accounts </div>
   <mat-card id="mainCard_matCard">
     <li *ngFor="let account of accounts">
@@ -22,5 +22,4 @@
       </mat-expansion-panel>
     </li>
   </mat-card>
-</div> -->
-<app-login></app-login>
+</div>  -->
