--- src/app/account/login/login.component.html
@@ -1,17 +1,0 @@
-<mat-card class="form-card">
-    <form class="form">
-        <mat-form-field  appearance="fill">
-        <mat-label>Email</mat-label>
-        <input type="email" matInput 
-                placeholder="Ex. pat@example.com">
-        </mat-form-field>
-        <mat-form-field  appearance="fill">
-            <mat-label>Password</mat-label>
-            <input type="password" matInput 
-                placeholder="********">
-        </mat-form-field>
-       <button mat-raised-button color="primary">Login</button>
-       <button mat-raised-button color="accent">Register</button>
-
-    </form>
-</mat-card>
