--- src/app/account/login/login.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-@Component({
-  selector: 'app-login',
-  templateUrl: './login.component.html',
-  styleUrls: ['./login.component.scss']
-})
-export class LoginComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
