+++ src/app/app.module.ts
@@ -7,21 +7,20 @@ import { AccountComponent } from './account/account.component';
 import { AccountService } from './account/service/account.service';
 import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import {MatExpansionModule} from '@angular/material/expansion';
-import {MatCardModule} from '@angular/material/card';
-import {MatInputModule} from '@angular/material/input';
-import {MatButtonModule} from '@angular/material/button';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatCardModule } from '@angular/material/card';
+import { MatInputModule } from '@angular/material/input';
+import { MatButtonModule } from '@angular/material/button';
 import { TransactionComponent } from './transaction/transaction.component';
 import { TransactionService } from './transaction/service/transaction.service';
 import { UserComponent } from './user/user.component';
 import { UserService } from './user/service/user.service';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginComponent } from './account/login/login.component';
-
-const appRoutes : Routes = [
-{path : 'transaction' ,component: TransactionComponent},
-{path : 'user' ,component: UserComponent},
-{path : 'login' ,component: LoginComponent},
+import { LoginComponent } from './user/login/login.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+const appRoutes: Routes = [
+  { path: 'transaction', component: TransactionComponent },
+  { path: 'login', component: LoginComponent },
 ]
 
 @NgModule({
@@ -41,10 +40,12 @@ const appRoutes : Routes = [
     MatCardModule,
     MatInputModule,
     MatButtonModule,
+    FormsModule,
+    ReactiveFormsModule,
     RouterModule.forRoot(appRoutes)
-  
+
   ],
-  providers: [AccountService,TransactionService,UserService],
+  providers: [AccountService, TransactionService, UserService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
