+++ src/app/user/user.component.html
@@ -1,1 +1,1 @@
-<p>user works!</p>
+<app-login></app-login>
